Refactor filenames to more closely match the domain
