+++ app-client/src/components/session/DayScheduleHeader.tsx
